flask_app/requirements.txt
Flask==1.1.1
gunicorn==20.0.4
Jinja2==2.11.1 
MarkupSafe==1.1.1 
Werkzeug==1.0.1 
click==7.1.1 
itsdangerous==1.1.0 
six==1.14.0
flask_wtf==0.14.3
WTForms==2.2.1
blinker==1.4
Flask-APScheduler==1.11.0
requests==2.24.0
Authlib==0.14.3
psycopg2-binary